lib/data_collection.py
# -*- coding: utf-8 -*-
import configparser
import requests

class DataCollector:
    def __init__(self):
        config = configparser.ConfigParser()
        config.read('config/config.ini')
        self.api_keys = config['API_KEYS']

    def collect_news_data(self, theme):
        # NewsAPIを使用してニュース記事を収集
        pass

    def collect_twitter_data(self, theme):
        # Twitter APIを使用してツイートを収集
        pass

    def collect_financial_data(self, companies):
        # Alpha Vantage API、Finnhub API、Polygon APIを使用して財務データを収集
        pass

    def collect_stock_price_data(self, companies):
        # Alpha Vantage API、Finnhub API、Polygon APIを使用して株価データを収集
        pass
